Moved config loading functions to helper. Some small edits.
